Extract tokenizer error treatment function in ExpressionEditorTextfield (#18706)
