+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/ExpressionEditorSuggestions/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ExpressionEditorSuggestions";
