+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/ExpressionEditorTextField.jsx
@@ -6,14 +6,11 @@ import { t } from "ttag";
 import _ from "underscore";
 import cx from "classnames";
 
+import { getTokenizerErrors } from "./tokenizerErrors";
+import { isExpression } from "metabase/lib/expressions";
 import { format } from "metabase/lib/expressions/format";
 import { processSource } from "metabase/lib/expressions/process";
-import {
-  tokenize,
-  countMatchingParentheses,
-  TOKEN,
-  OPERATOR as OP,
-} from "metabase/lib/expressions/tokenizer";
+import { tokenize } from "metabase/lib/expressions/tokenizer";
 import MetabaseSettings from "metabase/lib/settings";
 import colors from "metabase/lib/colors";
 
@@ -36,8 +33,6 @@ import Popover from "metabase/components/Popover";
 import ExplicitSize from "metabase/components/ExplicitSize";
 
 import TokenizedInput from "./TokenizedInput";
-
-import { getMBQLName, isExpression } from "metabase/lib/expressions";
 
 import ExpressionEditorSuggestions from "./ExpressionEditorSuggestions";
 
@@ -340,43 +335,13 @@ export default class ExpressionEditorTextfield extends React.Component {
     const showSuggestions =
       !hasSelection && !(isValid && isAtEnd && !endsWithWhitespace);
 
-    const { tokens, errors: tokenizerError } = tokenize(source);
-    const mismatchedParentheses = countMatchingParentheses(tokens);
-    const mismatchedError =
-      mismatchedParentheses === 1
-        ? t`Expecting a closing parenthesis`
-        : mismatchedParentheses > 1
-        ? t`Expecting ${mismatchedParentheses} closing parentheses`
-        : mismatchedParentheses === -1
-        ? t`Expecting an opening parenthesis`
-        : mismatchedParentheses < -1
-        ? t`Expecting ${-mismatchedParentheses} opening parentheses`
-        : null;
-    if (mismatchedError) {
-      tokenizerError.push({
-        message: mismatchedError,
-      });
-    }
-
-    for (let i = 0; i < tokens.length - 1; ++i) {
-      const token = tokens[i];
-      if (token.type === TOKEN.Identifier && source[token.start] !== "[") {
-        const functionName = source.slice(token.start, token.end);
-        if (getMBQLName(functionName)) {
-          const next = tokens[i + 1];
-          if (next.op !== OP.OpenParenthesis) {
-            tokenizerError.unshift({
-              message: t`Expecting an opening parenthesis after function ${functionName}`,
-            });
-          }
-        }
-      }
-    }
+    const { tokens, errors: tokenizerErrors } = tokenize(source);
+    getTokenizerErrors(source, tokens, tokenizerErrors);
 
     this.setState({
       source,
       expression,
-      tokenizerError,
+      tokenizerError: tokenizerErrors,
       compileError,
       displayError: null,
       suggestions: showSuggestions ? suggestions : [],
