+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/TokenizedInput/TokenizedExpression/TokenizedExpression.css
