+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/TokenizedInput/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TokenizedInput";
