+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorTextfield/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ExpressionEditorTextField";
